backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
chromadb
openai
tiktoken
pdfplumber
numpy
anthropic

